backend/tests/integration/test_registration.py
"""
Integration tests for user registration flow (User Story 1 - P1)
Tests the complete registration journey from API to database
"""
import pytest
from httpx import AsyncClient
from sqlalchemy.orm import Session


# Will be implemented after FastAPI app routes are created
# These tests should FAIL initially (TDD approach)

@pytest.mark.asyncio
async def test_register_success():
    """Test successful user registration (FR-001, FR-005)"""
    # TODO: Implement after /auth/register endpoint exists
    pass


@pytest.mark.asyncio
async def test_register_duplicate_email():
    """Test registration with existing email (FR-006)"""
    # TODO: Implement after /auth/register endpoint exists
    pass


@pytest.mark.asyncio
async def test_register_weak_password():
    """Test registration with weak password (FR-003)"""
    # TODO: Implement after /auth/register endpoint exists
    pass


@pytest.mark.asyncio
async def test_register_invalid_email():
    """Test registration with invalid email (FR-002)"""
    # TODO: Implement after /auth/register endpoint exists
    pass
