backend/requirements.txt
# Web Framework
fastapi[all]>=0.100.0
uvicorn[standard]>=0.23.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0

# Security
passlib[argon2]>=1.7.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
email-validator>=2.0.0

# Email
aiosmtplib>=2.0.0

# Configuration
python-dotenv>=1.0.0
